fix: move world belief to experimental package
