+++ dimos/experimental/world_belief/world_belief.py
@@ -25,18 +25,18 @@ import uuid
 import numpy as np
 from scipy.optimize import linear_sum_assignment
 
-from dimos.models.embedding.base import Embedding
-from dimos.msgs.geometry_msgs.Vector3 import Vector3
-from dimos.perception.experimental.world_belief.absence import (
+from dimos.experimental.world_belief.absence import (
     ABSENT,
     PRESENT,
     classify_visibility,
 )
-from dimos.perception.experimental.world_belief.identity_features import (
+from dimos.experimental.world_belief.identity_features import (
     add_diverse_embedding_view,
     gallery_cos,
     normalize_embedding,
 )
+from dimos.models.embedding.base import Embedding
+from dimos.msgs.geometry_msgs.Vector3 import Vector3
 from dimos.protocol.service.spec import BaseConfig
 from dimos.utils.logging_config import setup_logger
