+++ dimos/experimental/world_belief/xarm6_blueprint.py
@@ -24,17 +24,17 @@ import rerun.blueprint as rrb
 from dimos.agents.mcp.mcp_server import McpServer
 from dimos.constants import STATE_DIR
 from dimos.core.coordination.blueprints import autoconnect
+from dimos.experimental.world_belief.worldbelief_module import (
+    WorldBeliefModule,
+)
+from dimos.experimental.world_belief.worldbelief_recorder import (
+    WorldBeliefRecorder,
+)
 from dimos.hardware.sensors.camera.realsense.camera import RealSenseCamera
 from dimos.manipulation.manipulation_module import ManipulationModule
 from dimos.msgs.geometry_msgs.Quaternion import Quaternion
 from dimos.msgs.geometry_msgs.Transform import Transform
 from dimos.msgs.geometry_msgs.Vector3 import Vector3
-from dimos.perception.experimental.world_belief.worldbelief_module import (
-    WorldBeliefModule,
-)
-from dimos.perception.experimental.world_belief.worldbelief_recorder import (
-    WorldBeliefRecorder,
-)
 from dimos.robot.manipulators.common.blueprints import coordinator, trajectory_task
 from dimos.robot.manipulators.xarm.config import make_xarm6_model_config, xarm6_hardware
 from dimos.visualization.rerun.bridge import RerunBridgeModule
